Add a top-level fixture and override it in packages (#10)
